refactor(health): consolidate gRPC channel singleton with shared dashboard instance
The health app previously carried a local copy of GrpcChannelSingleton
under apps::health::services::grpc_channel as a workaround while PR #392
(crate::config::grpc_client::GrpcChannelSingleton) was in flight. PR #392
has been merged and the canonical singleton is registered globally via
its #[injectable_factory], so the duplicate factory is dead weight and
risks DI resolution surprises when both auto-register at the same scope.

Drop the local module, drop the empty services aggregator, and switch
both the healthz view and its integration test to consume the canonical
GrpcChannelSingleton from crate::config::grpc_client. The view now
reads grpc_channel.channel directly (the canonical singleton exposes
channel as a public field), removing the indirection through the
removed channel() accessor.

Closes #391
Refs #392
