+++ dashboard/src/apps/health.rs
@@ -8,7 +8,6 @@ use reinhardt::app_config;
 
 pub mod models;
 pub mod serializers;
-pub mod services;
 pub mod tests;
 pub mod urls;
 pub mod views;
