--- dashboard/src/apps/health/services.rs
@@ -1,5 +1,0 @@
-//! Services for the health app.
-
-pub mod grpc_channel;
-
-pub use grpc_channel::GrpcChannelSingleton;
